// metadata-calculator/src/model/HoumerSpeedInfoDTO.ts
export interface HoumerSpeedInfoDTO {
	houmerID: number;
	originDepartureDate: string;
	originLatitude: number;
	originLongitude: number;
	destiantionArrivalDate: string;
	destinationLatitude: number;
	destinarionLongitude: number;
	distanceInKM: number;
	transitTimeInHours: number;
	speed: string;
}